Monitor Interval type changed from uint16 to uint32 in itm-config.yang
as per RFC 5880 BFD Desired Min TX Interval is a 32 bit value specified in
microseconds .odl-interface.yang has uin32 as datatype for monitor
interval.This code changes will create sync in datatype for tunnel
monitoring interval.

+++ itm/itm-impl/src/main/java/org/opendaylight/genius/itm/cli/TepCommandHelper.java
@@ -422,7 +422,7 @@ public class TepCommandHelper {
         }
     }
 
-    public void showTeps(boolean monitorEnabled, int monitorInterval, CommandSession session) throws TepException {
+    public void showTeps(boolean monitorEnabled, long monitorInterval, CommandSession session) throws TepException {
         boolean flag = false;
         InstanceIdentifier<TransportZones> path = InstanceIdentifier.builder(TransportZones.class).build();
         Optional<TransportZones> tZonesOptional = ItmUtils.read(LogicalDatastoreType.CONFIGURATION, path, dataBroker);
@@ -855,7 +855,7 @@ public void showCache(String cacheName) {
         }
     }
 
-    public void configureTunnelMonitorInterval(int interval) {
+    public void configureTunnelMonitorInterval(long interval) {
         InstanceIdentifier<TunnelMonitorInterval> path = InstanceIdentifier.builder(TunnelMonitorInterval.class).build();
         Optional<TunnelMonitorInterval> storedTunnelMonitor = ItmUtils.read(LogicalDatastoreType.CONFIGURATION, path, dataBroker);
         if (!storedTunnelMonitor.isPresent() || storedTunnelMonitor.get().getInterval() != interval) {
